kube only deploy
when deploying with

kube_deploy=yes
swarm_deploy=no
mesos_deploy=no

some pieces of code try to work on swarm and mesos configurations and fail because the hosts file does not contain any mesos/swarm configurations.

updated the code to use the *_deploy flags before looking for the variables.